+++ tencentcloud/resource_tc_cam_group_membership.go
@@ -22,9 +22,9 @@ package tencentcloud
 
 import (
 	"context"
+	"fmt"
 	"log"
 	"strconv"
-	"time"
 
 	"github.com/hashicorp/terraform-plugin-sdk/helper/resource"
 	"github.com/hashicorp/terraform-plugin-sdk/helper/schema"
@@ -73,7 +73,28 @@ func resourceTencentCloudCamGroupMembershipCreate(d *schema.ResourceData, meta i
 		return err
 	}
 	d.SetId(groupId)
-	time.Sleep(3 * time.Second)
+
+	//get really instance then read
+	ctx := context.WithValue(context.TODO(), "logId", logId)
+
+	camService := CamService{
+		client: meta.(*TencentCloudClient).apiV3Conn,
+	}
+
+	err = resource.Retry(readRetryTimeout, func() *resource.RetryError {
+		instance, e := camService.DescribeGroupMembershipById(ctx, groupId)
+		if e != nil {
+			return retryError(e, "ResourceNotFound")
+		}
+		if len(instance) == 0 {
+			return resource.RetryableError(fmt.Errorf("creation not done"))
+		}
+		return nil
+	})
+	if err != nil {
+		log.Printf("[CRITAL]%s read CAM group membership failed, reason:%s\n", logId, err.Error())
+		return err
+	}
 
 	return resourceTencentCloudCamGroupMembershipRead(d, meta)
 }