+++ tencentcloud/resource_tc_cam_user_test.go
@@ -55,7 +55,7 @@ func TestAccTencentCloudCamUser_basic(t *testing.T) {
 				ResourceName:            "tencentcloud_cam_user.user_basic",
 				ImportState:             true,
 				ImportStateVerify:       true,
-				ImportStateVerifyIgnore: []string{"secret_key", "secret_id", "password", "need_reset_password", "use_api"},
+				ImportStateVerifyIgnore: []string{"secret_key", "secret_id", "password", "need_reset_password", "use_api", "force_delete"},
 			},
 		},
 	})
@@ -167,6 +167,7 @@ resource "tencentcloud_cam_user" "user_basic" {
   phone_num           = "12345678910"
   country_code        = "86"
   email               = "1234@qq.com"
+  force_delete		  = true
 }
 `
 
@@ -181,6 +182,7 @@ resource "tencentcloud_cam_user" "user_basic" {
   phone_num           = "13670093505"
   country_code        = "72"
   email               = "141515@qq.com"
+  force_delete		  = true
 }
 `
 const testAccCamUser_nilPassword = `
@@ -193,6 +195,7 @@ resource "tencentcloud_cam_user" "user_nil_password" {
 	phone_num           = "12345678910"
 	country_code        = "86"
 	email               = "141515@qq.com"
+    force_delete		= true
 }
 `
 const testAccCamUser_withoutKey = `
@@ -205,5 +208,6 @@ resource "tencentcloud_cam_user" "user_without_key" {
   phone_num           = "12345678910"
   country_code        = "86"
   email               = "141515@qq.com"
+  force_delete		  = true
 }
 `